+++ bundle/src/InternalBundle.php
@@ -6,6 +6,8 @@ use Hyvor\Internal\Auth\AuthFactory;
 use Hyvor\Internal\Auth\AuthInterface;
 use Hyvor\Internal\Billing\BillingFactory;
 use Hyvor\Internal\Billing\BillingInterface;
+use Hyvor\Internal\Bundle\Comms\Comms;
+use Hyvor\Internal\Bundle\Comms\CommsInterface;
 use Hyvor\Internal\Bundle\EventDispatcher\EventDispatcherCompilerPass;
 use Hyvor\Internal\InternalConfig;
 use Hyvor\Internal\SelfHosted\SelfHostedTelemetry;
@@ -63,7 +65,7 @@ class InternalBundle extends AbstractBundle
         $container->import('../config/services.php');
 
         $container->parameters()
-            ->set('internal.default_auth_method', 'oidc')
+            ->set('internal.default_deployment', 'on-prem')
             ->set('internal.default_instance', 'https://hyvor.com')
             ->set('internal.default_private_instance', null)
             ->set('internal.default_fake', false);
@@ -75,7 +77,7 @@ class InternalBundle extends AbstractBundle
                 '%env(APP_SECRET)%',
                 '%env(string:default::COMMS_KEY)%',
                 $config['component'],
-                '%env(default:internal.default_auth_method:AUTH_METHOD)%',
+                '%env(default:internal.default_deployment:DEPLOYMENT)%',
                 '%env(default:internal.default_instance:HYVOR_INSTANCE)%',
                 '%env(default:internal.default_private_instance:HYVOR_PRIVATE_INSTANCE)%',
                 '%env(bool:default:internal.default_fake:HYVOR_FAKE)%',
@@ -93,6 +95,10 @@ class InternalBundle extends AbstractBundle
             ->set(BillingInterface::class)
             ->public() // because this is not used from outside, so tests fail (inlined)
             ->factory([service(BillingFactory::class), 'create']);
+
+        $container
+            ->services()
+            ->alias(CommsInterface::class, Comms::class);
 
         // other services
         $container->services()->alias(SelfHostedTelemetryInterface::class, SelfHostedTelemetry::class);