+++ src/Auth/AuthUser.php
@@ -5,10 +5,10 @@ declare(strict_types=1);
 namespace Hyvor\Internal\Auth;
 
 use Hyvor\Internal\Bundle\Entity\OidcUser;
+use Symfony\Component\DependencyInjection\Attribute\Exclude;
 
 /**
  * @phpstan-type AuthUserArray array{
- *  current_organization?: array{id: int, name: string},
  *  id: int,
  *  username: string,
  *  name: string,
@@ -20,7 +20,6 @@ use Hyvor\Internal\Bundle\Entity\OidcUser;
  * }
  *
  * @phpstan-type AuthUserArrayPartial array{
- * current_organization?: array{id: int, name: string},
  * id?: int,
  * username?: string,
  * name?: string,
@@ -31,11 +30,11 @@ use Hyvor\Internal\Bundle\Entity\OidcUser;
  * website_url?: string,
  * }
  */
+#[Exclude]
 class AuthUser
 {
 
     final public function __construct(
-        public ?AuthCurrentOrganization $current_organization, // only set in AuthUser objects from ->check()
         public int $id,
         public string $username,
         public string $name,
@@ -54,10 +53,6 @@ class AuthUser
     public static function fromArray(array $data): static
     {
         return new static(
-            current_organization: isset($data['current_organization']) ? new AuthCurrentOrganization(
-                id: $data['current_organization']['id'],
-                name: $data['current_organization']['name'],
-            ) : null,
             id: $data['id'],
             username: $data['username'],
             name: $data['name'],
@@ -72,10 +67,6 @@ class AuthUser
     public static function fromOidcUser(OidcUser $oidcUser): static
     {
         return new static(
-            current_organization: new AuthCurrentOrganization(
-                id: 0,
-                name: 'Default',
-            ),
             id: $oidcUser->getId(),
             username: $oidcUser->getSub(),
             name: $oidcUser->getName(),
@@ -95,10 +86,6 @@ class AuthUser
     {
         /** @var AuthUserArray $user */
         $user = [
-            'current_organization' => $this->current_organization ? [
-                'id' => $this->current_organization->id,
-                'name' => $this->current_organization->name,
-            ] : null,
             'id' => $this->id,
             'username' => $this->username,
             'name' => $this->name,