+++ src/Billing/BillingInterface.php
@@ -3,20 +3,20 @@
 namespace Hyvor\Internal\Billing;
 
 use Hyvor\Internal\Billing\License\Resolved\ResolvedLicense;
+use Hyvor\Internal\Bundle\Comms\Exception\CommsApiFailedException;
 use Hyvor\Internal\Component\Component;
-use Hyvor\Internal\InternalApi\Exceptions\InternalApiCallFailedException;
 
 interface BillingInterface
 {
     /**
-     * @throws InternalApiCallFailedException
+     * @throws CommsApiFailedException
      */
     public function license(int $organizationId, ?Component $component = null): ResolvedLicense;
 
     /**
      * @param int[] $organizationIds
      * @return array<int, ResolvedLicense> orgId keyed licenses
-     * @throws InternalApiCallFailedException
+     * @throws CommsApiFailedException
      */
     public function licenses(array $organizationIds, ?Component $component = null): array;
 }