+++ src/Billing/Exception/LicenseOfCombinationNotFoundException.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Hyvor\Internal\Billing\Exception;
-
-class LicenseOfCombinationNotFoundException extends \Exception
-{
-
-}