+++ src/Billing/License/Resolved/ResolvedLicenseSubscription.php
@@ -2,7 +2,7 @@
 
 namespace Hyvor\Internal\Billing\License\Resolved;
 
-class ResolvedLicenseSubscription
+class ResolvedLicenseSubscription implements \JsonSerializable
 {
 
     public function __construct(
@@ -16,4 +16,16 @@ class ResolvedLicenseSubscription
     ) {
     }
 
+    public function jsonSerialize(): mixed
+    {
+        return [
+            'status' => $this->status,
+            'monthly_price' => $this->monthlyPrice,
+            'is_annual' => $this->isAnnual,
+            'plan' => $this->plan,
+            'plan_version' => $this->planVersion,
+            'plan_readable_name' => $this->planReadableName,
+            'cancel_at' => $this->cancelAt?->getTimestamp(),
+        ];
+    }
 }